Move BufferUtil.integerAsBuffer() to JSUtil and remove BufferUtil.
