+++ index.js
@@ -35,7 +35,6 @@ bsv.encoding.Varint = require('./lib/encoding/varint')
 
 // utilities
 bsv.util = {}
-bsv.util.buffer = require('./lib/util/buffer')
 bsv.util.js = require('./lib/util/js')
 bsv.util.preconditions = require('./lib/util/preconditions')
