+++ lib/hdprivatekey.js
@@ -16,7 +16,6 @@ var Random = require('./crypto/random')
 
 var errors = require('./errors')
 var hdErrors = errors.HDPrivateKey
-var BufferUtil = require('./util/buffer')
 var JSUtil = require('./util/js')
 
 var MINIMUM_ENTROPY_BITS = 128
@@ -232,7 +231,7 @@ HDPrivateKey.prototype._deriveWithNumber = function (index, hardened, nonComplia
     index += HDPrivateKey.Hardened
   }
 
-  var indexBuffer = BufferUtil.integerAsBuffer(index)
+  var indexBuffer = JSUtil.integerAsBuffer(index)
   var data
   if (hardened && nonCompliant) {
     // The private key serialization in this case will not be exactly 32 bytes and can be
@@ -362,13 +361,13 @@ HDPrivateKey.prototype._buildFromJSON = function (arg) {
 HDPrivateKey.prototype._buildFromObject = function (arg) {
   // TODO: Type validation
   var buffers = {
-    version: arg.network ? BufferUtil.integerAsBuffer(Network.get(arg.network).xprivkey) : arg.version,
+    version: arg.network ? JSUtil.integerAsBuffer(Network.get(arg.network).xprivkey) : arg.version,
     depth: _.isNumber(arg.depth) ? Buffer.from([arg.depth & 0xff]) : arg.depth,
-    parentFingerPrint: _.isNumber(arg.parentFingerPrint) ? BufferUtil.integerAsBuffer(arg.parentFingerPrint) : arg.parentFingerPrint,
-    childIndex: _.isNumber(arg.childIndex) ? BufferUtil.integerAsBuffer(arg.childIndex) : arg.childIndex,
+    parentFingerPrint: _.isNumber(arg.parentFingerPrint) ? JSUtil.integerAsBuffer(arg.parentFingerPrint) : arg.parentFingerPrint,
+    childIndex: _.isNumber(arg.childIndex) ? JSUtil.integerAsBuffer(arg.childIndex) : arg.childIndex,
     chainCode: _.isString(arg.chainCode) ? Buffer.from(arg.chainCode, 'hex') : arg.chainCode,
     privateKey: (_.isString(arg.privateKey) && JSUtil.isHexa(arg.privateKey)) ? Buffer.from(arg.privateKey, 'hex') : arg.privateKey,
-    checksum: arg.checksum ? (arg.checksum.length ? arg.checksum : BufferUtil.integerAsBuffer(arg.checksum)) : undefined
+    checksum: arg.checksum ? (arg.checksum.length ? arg.checksum : JSUtil.integerAsBuffer(arg.checksum)) : undefined
   }
   return this._buildFromBuffers(buffers)
 }
