+++ lib/hdpublickey.js
@@ -18,7 +18,6 @@ var hdErrors = bsvErrors.HDPublicKey
 var assert = require('assert')
 
 var JSUtil = require('./util/js')
-var BufferUtil = require('./util/buffer')
 
 /**
  * The representation of an hierarchically derived public key.
@@ -159,7 +158,7 @@ HDPublicKey.prototype._deriveWithNumber = function (index, hardened) {
     throw new hdErrors.InvalidPath(index)
   }
 
-  var indexBuffer = BufferUtil.integerAsBuffer(index)
+  var indexBuffer = JSUtil.integerAsBuffer(index)
   var data = Buffer.concat([this.publicKey.toBuffer(), indexBuffer])
   var hash = Hash.sha512hmac(data, this._buffers.chainCode)
   var leftPart = BN.fromBuffer(hash.slice(0, 32), { size: 32 })
@@ -265,7 +264,7 @@ HDPublicKey.prototype._buildFromPrivate = function (arg) {
   var args = _.clone(arg._buffers)
   var point = Point.getG().mul(BN.fromBuffer(args.privateKey))
   args.publicKey = Point.pointToCompressed(point)
-  args.version = BufferUtil.integerAsBuffer(Network.get(args.version.readUInt32BE(0)).xpubkey)
+  args.version = JSUtil.integerAsBuffer(Network.get(args.version.readUInt32BE(0)).xpubkey)
   args.privateKey = undefined
   args.checksum = undefined
   args.xprivkey = undefined
@@ -275,14 +274,14 @@ HDPublicKey.prototype._buildFromPrivate = function (arg) {
 HDPublicKey.prototype._buildFromObject = function (arg) {
   // TODO: Type validation
   var buffers = {
-    version: arg.network ? BufferUtil.integerAsBuffer(Network.get(arg.network).xpubkey) : arg.version,
+    version: arg.network ? JSUtil.integerAsBuffer(Network.get(arg.network).xpubkey) : arg.version,
     depth: _.isNumber(arg.depth) ? Buffer.from([arg.depth & 0xff]) : arg.depth,
-    parentFingerPrint: _.isNumber(arg.parentFingerPrint) ? BufferUtil.integerAsBuffer(arg.parentFingerPrint) : arg.parentFingerPrint,
-    childIndex: _.isNumber(arg.childIndex) ? BufferUtil.integerAsBuffer(arg.childIndex) : arg.childIndex,
+    parentFingerPrint: _.isNumber(arg.parentFingerPrint) ? JSUtil.integerAsBuffer(arg.parentFingerPrint) : arg.parentFingerPrint,
+    childIndex: _.isNumber(arg.childIndex) ? JSUtil.integerAsBuffer(arg.childIndex) : arg.childIndex,
     chainCode: _.isString(arg.chainCode) ? Buffer.from(arg.chainCode, 'hex') : arg.chainCode,
     publicKey: _.isString(arg.publicKey) ? Buffer.from(arg.publicKey, 'hex')
       : Buffer.isBuffer(arg.publicKey) ? arg.publicKey : arg.publicKey.toBuffer(),
-    checksum: _.isNumber(arg.checksum) ? BufferUtil.integerAsBuffer(arg.checksum) : arg.checksum
+    checksum: _.isNumber(arg.checksum) ? JSUtil.integerAsBuffer(arg.checksum) : arg.checksum
   }
   return this._buildFromBuffers(buffers)
 }
