+++ lib/networks.js
@@ -1,7 +1,6 @@
 'use strict'
 var _ = require('./util/_')
 
-var BufferUtil = require('./util/buffer')
 var JSUtil = require('./util/js')
 var networks = []
 var networkMaps = {}
@@ -102,7 +101,7 @@ function addNetwork (data) {
 
   if (data.networkMagic) {
     _.extend(network, {
-      networkMagic: BufferUtil.integerAsBuffer(data.networkMagic)
+      networkMagic: JSUtil.integerAsBuffer(data.networkMagic)
     })
   }
 
@@ -287,11 +286,11 @@ Object.defineProperty(testnet, 'networkMagic', {
   configurable: false,
   get: function () {
     if (this.regtestEnabled) {
-      return BufferUtil.integerAsBuffer(REGTEST.NETWORK_MAGIC)
+      return JSUtil.integerAsBuffer(REGTEST.NETWORK_MAGIC)
     } else if (this.stnEnabled) {
-      return BufferUtil.integerAsBuffer(STN.NETWORK_MAGIC)
+      return JSUtil.integerAsBuffer(STN.NETWORK_MAGIC)
     } else {
-      return BufferUtil.integerAsBuffer(TESTNET.NETWORK_MAGIC)
+      return JSUtil.integerAsBuffer(TESTNET.NETWORK_MAGIC)
     }
   }
 })
