--- lib/util/buffer.js
@@ -1,22 +1,0 @@
-'use strict'
-
-var $ = require('./preconditions')
-
-module.exports = {
-  /**
-   * Transform a 4-byte integer into a Buffer of length 4.
-   *
-   * @param {number} integer
-   * @return {Buffer}
-   */
-  integerAsBuffer: function integerAsBuffer (integer) {
-    $.checkArgumentType(integer, 'number', 'integer')
-    var bytes = []
-    bytes.push((integer >> 24) & 0xff)
-    bytes.push((integer >> 16) & 0xff)
-    bytes.push((integer >> 8) & 0xff)
-    bytes.push(integer & 0xff)
-    return Buffer.from(bytes)
-  }
-
-}
