+++ lib/util/js.js
@@ -1,6 +1,7 @@
 'use strict'
 
 var _ = require('../util/_')
+var $ = require('./preconditions')
 
 /**
  * Determines whether a string contains only hexadecimal values
@@ -80,5 +81,21 @@ module.exports = {
       isFinite(value) &&
       Math.floor(value) === value &&
       value >= 0
+  },
+
+  /**
+   * Transform a 4-byte integer into a Buffer of length 4.
+   *
+   * @param {number} integer
+   * @return {Buffer}
+   */
+  integerAsBuffer: function integerAsBuffer (integer) {
+    $.checkArgumentType(integer, 'number', 'integer')
+    var bytes = []
+    bytes.push((integer >> 24) & 0xff)
+    bytes.push((integer >> 16) & 0xff)
+    bytes.push((integer >> 8) & 0xff)
+    bytes.push(integer & 0xff)
+    return Buffer.from(bytes)
   }
 }
